+++ src/Definitions/Carrier.php
@@ -4,6 +4,10 @@ declare(strict_types=1);
 
 namespace Inspirum\Balikobot\Definitions;
 
+use function array_filter;
+use function array_values;
+use function in_array;
+
 final class Carrier extends BaseEnum
 {
     /**
@@ -63,6 +67,10 @@ final class Carrier extends BaseEnum
 
     /**
      * WE DO - Uloženka
+     *
+     * @deprecated Replaced by One by Allegro
+     *
+     * @see self::ONEBYALLEGRO
      */
     public const ULOZENKA = 'ulozenka';
 
@@ -143,6 +151,8 @@ final class Carrier extends BaseEnum
 
     /**
      * 123kurier
+     *
+     * @deprecated
      */
     public const KURIER = 'kurier';
 
@@ -183,6 +193,8 @@ final class Carrier extends BaseEnum
 
     /**
      * Quality Delivery Logistics
+     *
+     * @deprecated
      */
     public const QDL = 'qdl';
 
@@ -190,4 +202,21 @@ final class Carrier extends BaseEnum
      * Inpost
      */
     public const INPOST = 'inpost';
+
+    /**
+     * One by Allegro
+     */
+    public const ONEBYALLEGRO = 'onebyallegro';
+
+    /**
+     * @return list<string>
+     */
+    public static function getAll(): array
+    {
+        return array_values(array_filter(parent::getAll(), static fn (string $value): bool => !in_array($value, [
+            self::ULOZENKA,
+            self::QDL,
+            self::KURIER,
+        ])));
+    }
 }